bump version to 2.0 and change URL to fork
The upstream pyelliptic has been deprecated by the author at 1.5.8
and ECC API has been removed.

This version is a fork of the pyelliptic from the BitMessage source
tree: https://github.com/Bitmessage/PyBitmessage
The BitMessage fork forked from v1.3 of upstream pyelliptic.

The source in this version 2.0 is the BitMessage fork commits extracted
from BitMessage source repository and applied to pyelliptic v1.3 (i.e.
the base of the fork). Also, on top of those commit one extra fix is
added that is not (yet) in the BitMessage source (see
928e8e4421ed2d655926ab06adc11e0ac46b1ba0 in this repo).